+++ controller/tracepipeline/config_builder.go
@@ -2,6 +2,7 @@ package tracepipeline
 
 import (
 	"fmt"
+	"strings"
 
 	"github.com/kyma-project/kyma/components/telemetry-operator/apis/telemetry/v1alpha1"
 )
@@ -174,7 +175,7 @@ func makeHeaders(output v1alpha1.TracePipelineOutput) map[string]string {
 		headers["Authorization"] = fmt.Sprintf("${%s}", basicAuthHeaderVariable)
 	}
 	for _, header := range output.Otlp.Headers {
-		headers[header.Name] = fmt.Sprintf("${HEADER_%s}", envvar.MakeEnvVarCompliant(header.Name))
+		headers[header.Name] = fmt.Sprintf("${HEADER_%s}", MakeEnvVarCompliant(header.Name))
 	}
 	return headers
 }
@@ -346,4 +347,12 @@ func makeOtelCollectorConfig(output v1alpha1.TracePipelineOutput, isInsecureOutp
 		Extensions: extensionConfig,
 		Service:    serviceConfig,
 	}
+}
+
+func MakeEnvVarCompliant(input string) string {
+	result := input
+	result = strings.ToUpper(result)
+	result = strings.Replace(result, ".", "_", -1)
+	result = strings.Replace(result, "-", "_", -1)
+	return result
 }
